requirements.txt
flask
ultralytics
opencv-python
flask-cors
Pillow